Update double PotenciaRecursiva.java

+++ src/main/java/com/example/PotenciaRecursiva.java
@@ -2,15 +2,16 @@ package com.example;
 
 public class PotenciaRecursiva {
        
-    public static int multiplicar(int a, int b) {
+    public static double multiplicar(int a, int b) {
         if (b == 0) return 0;
         return a + multiplicar(a, b - 1);
     }
 
     
-    public static int potencia(int base, int exponente) {
+    public static double potencia(int base, int exponente) {
         if (exponente == 0) return 1;
         return multiplicar(base, potencia(base, exponente - 1));
     }
 }
+
